poc1/requirements.txt
Django==2.1
djangorestframework==3.8.2
django-rest-swagger==2.2.0
gunicorn==19.9.0
